front: reorganize rolling stock module
