+++ front/src/modules/rollingStock/components/RollingStockSelector/index.ts
@@ -1,9 +1,7 @@
 /* eslint-disable import/prefer-default-export */
 export { default as RollingStockSelector } from 'modules/rollingStock/components/RollingStockSelector/WithRollingStockSelector';
-export { default as RollingStock2Img } from 'modules/rollingStock/components/RollingStockSelector/RollingStock2Img';
 export { default as RollingStockCard } from 'modules/rollingStock/components/RollingStockSelector/RollingStockCard';
 export { default as RollingStockCardButtons } from 'modules/rollingStock/components/RollingStockSelector/RollingStockCardButtons';
 export { default as RollingStockCardDetail } from 'modules/rollingStock/components/RollingStockSelector/RollingStockCardDetail';
-export { default as RollingStockCurve } from 'modules/rollingStock/components/RollingStockSelector/RollingStockCurve';
 export { default as RollingStockModal } from 'modules/rollingStock/components/RollingStockSelector/RollingStockModal';
 export { default as SearchRollingStock } from 'modules/rollingStock/components/RollingStockSelector/SearchRollingStock';
