+++ front/src/modules/trainschedule/components/Timetable/TimetableTrainCard.tsx
@@ -12,7 +12,7 @@ import cx from 'classnames';
 import { InvalidTrainValues, osrdEditoastApi } from 'common/api/osrdEditoastApi';
 
 import { ScheduledTrain } from 'reducers/osrdsimulation/types';
-import { RollingStock2Img } from 'modules/rollingStock/components/RollingStockSelector';
+import RollingStock2Img from 'modules/rollingStock/components/RollingStock2Img';
 import { updateTrainScheduleIDsToModify } from 'reducers/osrdconf';
 import { useDispatch } from 'react-redux';
 import { jouleToKwh, mToKmOneDecimal } from 'utils/physics';
